Using all doubles

+++ src/complex.h
@@ -12,7 +12,7 @@ int complex_add(_complex *self, const _complex *other);
 int complex_mult(_complex *self, const _complex *other);
 //int complex_div(_complex *self, _complex *other);
 int strtoc(_complex *self, char* str);
-float complex_abs(const _complex *self);
+double complex_abs(const _complex *self);
 
 double complex_getX(const _complex *self);
 double complex_getY(const _complex *self);
